+++ src-v4/Corvus.Json.CodeGeneration.CSharp/Corvus.Json.CodeGeneration/CSharp/CodeGeneratorExtensions.cs
@@ -280,7 +280,7 @@ internal static partial class CodeGeneratorExtensions
             /// </summary>
             """)
             .AppendIndent("public static string SchemaLocation { get; } = ")
-            .Append(SymbolDisplay.FormatLiteral(typeDeclaration.RelativeSchemaLocation, true))
+            .Append(SymbolDisplay.FormatLiteral(typeDeclaration.LocatedSchema.RootDocumentPointer, true))
             .AppendLine(";");
     }